doc: add doxygen and format lexer_tok and lexer_src, create lexer_tok_utils
