+++ inc/lexer_tok_utils.h
@@ -1,35 +1,24 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   lexer_tok.h                                        :+:      :+:    :+:   */
+/*   lexer_tok_utils.h                                  :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: sqiu <sqiu@student.42vienna.com>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/07/02 18:52:14 by sqiu              #+#    #+#             */
-/*   Updated: 2023/07/12 22:42:45 by sqiu             ###   ########.fr       */
+/*   Created: 2023/07/12 21:47:14 by sqiu              #+#    #+#             */
+/*   Updated: 2023/07/12 22:04:57 by sqiu             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#ifndef LEXER_TOK_H
-# define LEXER_TOK_H
+#ifndef LEXER_TOK_UTILS_H
+# define LEXER_TOK_UTILS_H
 
 /* ====== Includes ====== */
 
-//# include "minishell.h"
-# include "lexer_src.h"
-# include "lexer_utils.h"
-# include "lexer_tok_utils.h"
-# include "minishell_error.h"
-
-/* ====== Definitions ====== */
-
-
-
 /* ====== Functions ====== */
 
-t_error	ft_tokenise(t_src *src, t_tok *token);
-t_error	ft_create_tok(t_tok *token, char *s);
-t_error	ft_partition(t_src *src, char *buf, int *buf_pos, int *buf_size);
-t_error	ft_init_partition(t_src *src, char *c);
+t_error	ft_init_tok(t_data *data);
+void	ft_free_tok(t_tok *token);
+t_error	ft_add_to_buf(char c, char *buf, int *buf_size, int *buf_pos);
 
 #endif
